requirements.txt
django
django-cities-light
django-smart-selects
requests
django-rest-framework
